training/requirements.txt
# Core Data Science (Pinned for Stability)
numpy==1.24.4
pandas==1.5.3
scikit-learn==1.2.2

# Deep Learning & ML
xgboost==1.7.6
tensorflow==2.11.1
mlxtend==0.22.0
catboost==1.2.0

# Validation & Orchestration
deepchecks==0.18.1
anywidget==0.9.3
prefect==2.19.4
griffe==0.49.0
pydantic==1.10.14

# Infrastructure & Utilities
sqlalchemy==1.4.49
psycopg2-binary==2.9.9
huggingface_hub==0.20.3
joblib==1.2.0
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0